+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The sections should follow the order `Added`, `Changed`, `Fixed`, and `Removed`.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- Rasterizer no longer uses associated `Err` type and now has common type for `Error`
+- If the glyph you've requested is missing, the resulted missing glyph char will be in `Error::MissingGlyph`
+
 ## 0.1.1
 
 ### Changed